move product-list ssr to its own directory
https://coveord.atlassian.net/browse/KIT-3488
